Move DTOs to application layer
We will use DTOs to provide simple data structures without logic to
the presentation layer. The domain layer will be focused only on
business logic and domain types.
