+++ src/main/java/com/ite/itea/application/dto/ItemDto.java
@@ -1,4 +1,4 @@
-package com.ite.itea.domain.dto;
+package com.ite.itea.application.dto;
 
 public abstract class ItemDto {
