+++ src/main/java/com/ite/itea/application/dto/TablesDto.java
@@ -1,4 +1,4 @@
-package com.ite.itea.domain.dto;
+package com.ite.itea.application.dto;
 
 import com.ite.itea.domain.request.ItemNameRequest;
