+++ src/main/java/com/ite/itea/domain/user/UserManagement.java
@@ -1,7 +1,7 @@
 package com.ite.itea.domain.user;
 
 import com.ite.itea.domain.Printer;
-import com.ite.itea.domain.dto.UserDto;
+import com.ite.itea.application.dto.UserDto;
 import com.ite.itea.persistence.UserRepository;
 
 import java.util.List;
